change backend

+++ testFront/contracts/TokenForDeposit.sol
@@ -5,21 +5,25 @@ import "hardhat/console.sol";
 
 contract TokenForDeposit {
     struct user {
-        uint deposit;
-        uint id;
+        uint256 deposit;
+        uint256 id;
     }
 
-    uint public userId;
-    event AddDeposit(address _address, uint _current_amount, uint _total_amount);
+    uint256 public userId;
+    event AddDeposit(
+        address _address,
+        uint256 _current_amount,
+        uint256 _total_amount
+    );
 
     mapping(address => user) public Users;
-    mapping(uint => address) public UserAddressById;
+    mapping(uint256 => address) public UserAddressById;
 
-    function getUserDeposit(address _address) public view returns(uint){
+    function getUserDeposit(address _address) public view returns (uint256) {
         return Users[_address].deposit;
     }
 
-    function addDeposit(uint _deposit) public {
+    function addDeposit(uint256 _deposit) public {
         userId += 1;
         Users[msg.sender].deposit += _deposit;
         Users[msg.sender].id = userId;
